Move update faq's  service/repo/model to features's updateFaq
