+++ app/faq-service/src/main/java/com/corporatebanking/faq/features/updateFaq/model/UpdateFaqModel.java
@@ -1,4 +1,4 @@
-package com.corporatebanking.faq.model;
+package com.corporatebanking.faq.features.updateFaq.model;
 
 import java.time.LocalDateTime;
