+++ app/faq-service/src/main/java/com/corporatebanking/faq/features/updateFaq/repository/UpdateFaqCategoryRepository.java
@@ -1,0 +1,11 @@
+package com.corporatebanking.faq.features.updateFaq.repository;
+
+import com.corporatebanking.faq.features.updateFaq.model.UpdateFaqCategoryModel;
+
+import java.util.Optional;
+
+public interface UpdateFaqCategoryRepository {
+
+  public Optional<UpdateFaqCategoryModel> findById(int id);
+   
+}
