--- app/faq-service/src/main/java/com/corporatebanking/faq/repository/UpdateFaqCategoryRepository.java
@@ -1,11 +1,0 @@
-package com.corporatebanking.faq.repository;
-
-import com.corporatebanking.faq.model.UpdateFaqCategoryModel;
-
-import java.util.Optional;
-
-public interface UpdateFaqCategoryRepository {
-
-  public Optional<UpdateFaqCategoryModel> findById(int id);
-   
-}
